+++ ImageFeed/Auth/AuthViewController.swift
@@ -11,9 +11,11 @@ final class AuthViewController: UIViewController {
 
     override func prepare(for segue: UIStoryboardSegue, sender: Any?) {
         if segue.identifier == ShowWebViewSegueIdentifier {
+            
             guard
                 let webViewViewController = segue.destination as? WebViewViewController
             else { fatalError("Failed to prepare for \(ShowWebViewSegueIdentifier)") }
+            
             webViewViewController.delegate = self
         } else {
             super.prepare(for: segue, sender: sender)
