+++ ImageFeed/Constants.swift
@@ -7,8 +7,8 @@
 
 import Foundation
 
-let AccessKey = "h9u9IDtscwhLMDSevmIPK0YXBMwvUZVfSCEbykIdB1E"
-let SecretKey = "Ip1ortflUIXhh_r0e1Qs37a01SzAor5fMsr8StqFqm0"
-let RedirectURI = "urn:ietf:wg:oauth:2.0:oob"
-let AccessScope = "public+read_user+write_likes"
-let DefaultBaseURL = URL(string: "https://api.unsplash.com")!
+let accessKey = "h9u9IDtscwhLMDSevmIPK0YXBMwvUZVfSCEbykIdB1E"
+let secretKey = "Ip1ortflUIXhh_r0e1Qs37a01SzAor5fMsr8StqFqm0"
+let redirectURI = "urn:ietf:wg:oauth:2.0:oob"
+let accessScope = "public+read_user+write_likes"
+let defaultBaseURL = URL(string: "https://api.unsplash.com")!
