+++ ImageFeed/Network/NetworkService.swift
@@ -1,97 +1,15 @@
 import UIKit
 import Foundation
 
-
-final class OAuth2TokenStorage {
-    private let userDefaults = UserDefaults.standard
-    private enum Keys: String, CodingKey {
-        case bearerToken
-    }
-    
-    var token: String? {
-        get {
-            return userDefaults.string(forKey: Keys.bearerToken.rawValue)
-        }
-        set {
-            userDefaults.set(newValue, forKey: Keys.bearerToken.rawValue)
-            print(userDefaults.string(forKey: Keys.bearerToken.rawValue))
-        }
-    }
-}
-
-// MARK: - Слой доступа к сервису (Service Access Layer)
-
-final class OAuth2Service {
-    static let shared = OAuth2Service()
-    
-    private let urlSession = URLSession.shared
-    
-    private (set) var authToken: String? {
-        get {
-            return OAuth2TokenStorage().token
-        }
-        set {
-            OAuth2TokenStorage().token = newValue
-        }
-    }
-
-    
-    func fetchOAuthToken(_ code: String, completion: @escaping (Result<String, Error>) -> Void) {
-        let request = authTokenRequest(code: code)
-        let task = object(for: request) { [weak self] result in
-            guard let self = self else { return }
-            switch result {
-            case .success(let body):
-                let authToken = body.accessToken
-                self.authToken = authToken
-                completion(.success(authToken))
-            case .failure(let error):
-                completion(.failure(error))
-            }
-        }
-        task.resume()
-    }
-}
-
-extension OAuth2Service {
-    private func object(
-        for request: URLRequest,
-        completion: @escaping (Result<OAuthTokenResponseBody, Error>) -> Void
-    ) -> URLSessionTask {
-        let decoder = JSONDecoder()
-        return urlSession.data(for: request) { (result: Result<Data, Error>) in
-            let response = result.flatMap { data -> Result<OAuthTokenResponseBody, Error> in
-                Result { try decoder.decode(OAuthTokenResponseBody.self, from: data) }
-            }
-            completion(response)
-        }
-    }
-
-    
-    private func authTokenRequest(code: String) -> URLRequest {
-            URLRequest.makeHTTPRequest(
-                path: "/oauth/token"
-                + "?client_id=\(AccessKey)"
-                + "&&client_secret=\(SecretKey)"
-                + "&&redirect_uri=\(RedirectURI)"
-                + "&&code=\(code)"
-                + "&&grant_type=authorization_code",
-                httpMethod: "POST",
-                baseURL: URL(string: "https://unsplash.com")!
-            )
-    }
-    
-}
-
 // MARK: - HTTP Requset
 
-//fileprivate let DefaultBaseURL = URL(string: "https://api.unsplash.com")!
+//fileprivate let defaultBaseURL = URL(string: "https://api.unsplash.com")!
 
 extension URLRequest {
     static func makeHTTPRequest(
         path: String,
         httpMethod: String,
-        baseURL: URL = DefaultBaseURL
+        baseURL: URL = defaultBaseURL
     ) -> URLRequest {
         var request = URLRequest(url: URL(string: path, relativeTo: baseURL)!)
         request.httpMethod = httpMethod
